Executor-based frida mode
