+++ include/fuzzuf/cli/fuzzer/afl/build_afl_fuzzer_from_args.hpp
@@ -97,22 +97,9 @@ std::unique_ptr<TFuzzer> BuildAFLFuzzerFromArgs(
         fuzzuf::cli::fuzzer::afl::usage(fuzzer_args.global_options_description);
     }
 
-    u32 extra_mem = 0;
     if (afl_options.frida_mode) {
-        setenv("__AFL_DEFER_FORKSRV", "1", 1);
-        fs::path frida_bin = fs::path(fuzzer_args.argv[0]).parent_path() / "afl-frida-trace.so";
-
-        struct stat statbuf;
-        if ((stat(frida_bin.c_str(), &statbuf)) == -1) {
-            std::cerr << cLRD <<
-                "[-] File afl-frida-trace.so not found\n" <<
-                "    Build one first with cmake where -DENABLE_FRIDA_TRACE=1" <<
-                cRST << std::endl;
-        }
-        // Need to add the size of the library
-        extra_mem += statbuf.st_size;
-
-        setenv("LD_PRELOAD", frida_bin.c_str(), 1);
+        // One of executor classes will handle frida mode later
+        setenv("FUZZUF_FRIDA_MODE", "1", 1);
     }
 
     PutArgs put(pargs);
@@ -143,7 +130,7 @@ std::unique_ptr<TFuzzer> BuildAFLFuzzerFromArgs(
                         global_options.in_dir,
                         global_options.out_dir,
                         global_options.exec_timelimit_ms.value_or(GetExecTimeout<AFLTag>()),
-                        global_options.exec_memlimit.value_or(GetMemLimit<AFLTag>()) + extra_mem,
+                        global_options.exec_memlimit.value_or(GetMemLimit<AFLTag>()),
                         afl_options.forksrv,
                         /* dumb_mode */ false,  // FIXME: add dumb_mode
                         Util::CPUID_BIND_WHICHEVER
